fix:案件管理

+++ src/main/java/com/ycw/fxq/bean/User.java
@@ -6,5 +6,6 @@ import lombok.EqualsAndHashCode;
 @Data
 @EqualsAndHashCode(callSuper=false)
 public class User {
-    private Integer id;
+    // TODO
+    private Long id;
 }
